Split EnvelopeSource into LogEnvelopeSource and SessionEnvelopeSource
