--- embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/capture/envelope/EnvelopeSource.kt
@@ -1,8 +1,0 @@
-package io.embrace.android.embracesdk.capture.envelope
-
-import io.embrace.android.embracesdk.internal.payload.Envelope
-import io.embrace.android.embracesdk.session.orchestrator.SessionSnapshotType
-
-internal fun interface EnvelopeSource<T> {
-    fun getEnvelope(endType: SessionSnapshotType): Envelope<T>
-}
