+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/capture/envelope/log/LogEnvelopeSourceImpl.kt
@@ -1,19 +1,16 @@
-package io.embrace.android.embracesdk.capture.envelope
+package io.embrace.android.embracesdk.capture.envelope.log
 
-import io.embrace.android.embracesdk.capture.envelope.log.LogPayloadSource
 import io.embrace.android.embracesdk.capture.envelope.metadata.EnvelopeMetadataSource
 import io.embrace.android.embracesdk.capture.envelope.resource.EnvelopeResourceSource
 import io.embrace.android.embracesdk.internal.payload.Envelope
-import io.embrace.android.embracesdk.internal.payload.LogPayload
-import io.embrace.android.embracesdk.session.orchestrator.SessionSnapshotType
 
-internal class LogEnvelopeSource(
+internal class LogEnvelopeSourceImpl(
     private val metadataSource: EnvelopeMetadataSource,
     private val resourceSource: EnvelopeResourceSource,
     private val logPayloadSource: LogPayloadSource,
-) : EnvelopeSource<LogPayload> {
+) : LogEnvelopeSource {
 
-    override fun getEnvelope(endType: SessionSnapshotType) = Envelope(
+    override fun getEnvelope() = Envelope(
         resourceSource.getEnvelopeResource(),
         metadataSource.getEnvelopeMetadata(),
         null,
