+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/capture/envelope/session/SessionEnvelopeSource.kt
@@ -1,0 +1,9 @@
+package io.embrace.android.embracesdk.capture.envelope.session
+
+import io.embrace.android.embracesdk.internal.payload.Envelope
+import io.embrace.android.embracesdk.internal.payload.SessionPayload
+import io.embrace.android.embracesdk.session.orchestrator.SessionSnapshotType
+
+internal fun interface SessionEnvelopeSource {
+    fun getEnvelope(endType: SessionSnapshotType): Envelope<SessionPayload>
+}
