+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/capture/envelope/session/SessionEnvelopeSourceImpl.kt
@@ -1,17 +1,16 @@
-package io.embrace.android.embracesdk.capture.envelope
+package io.embrace.android.embracesdk.capture.envelope.session
 
 import io.embrace.android.embracesdk.capture.envelope.metadata.EnvelopeMetadataSource
 import io.embrace.android.embracesdk.capture.envelope.resource.EnvelopeResourceSource
-import io.embrace.android.embracesdk.capture.envelope.session.SessionPayloadSource
 import io.embrace.android.embracesdk.internal.payload.Envelope
 import io.embrace.android.embracesdk.internal.payload.SessionPayload
 import io.embrace.android.embracesdk.session.orchestrator.SessionSnapshotType
 
-internal class SessionEnvelopeSource(
+internal class SessionEnvelopeSourceImpl(
     private val metadataSource: EnvelopeMetadataSource,
     private val resourceSource: EnvelopeResourceSource,
     private val sessionPayloadSource: SessionPayloadSource,
-) : EnvelopeSource<SessionPayload> {
+) : SessionEnvelopeSource {
 
     override fun getEnvelope(endType: SessionSnapshotType): Envelope<SessionPayload> {
         return Envelope(
