+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/capture/envelope/log/LogEnvelopeSourceImplTest.kt
@@ -1,26 +1,25 @@
-package io.embrace.android.embracesdk.capture.envelope
+package io.embrace.android.embracesdk.capture.envelope.log
 
 import io.embrace.android.embracesdk.fakes.FakeEnvelopeMetadataSource
 import io.embrace.android.embracesdk.fakes.FakeEnvelopeResourceSource
 import io.embrace.android.embracesdk.fakes.FakeLogPayloadSource
-import io.embrace.android.embracesdk.session.orchestrator.SessionSnapshotType
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
 
-internal class LogEnvelopeSourceTest {
+internal class LogEnvelopeSourceImplTest {
 
     @Test
     fun getEnvelope() {
         val metadataSource = FakeEnvelopeMetadataSource()
         val resourceSource = FakeEnvelopeResourceSource()
         val logSource = FakeLogPayloadSource()
-        val source = LogEnvelopeSource(
+        val source = LogEnvelopeSourceImpl(
             metadataSource,
             resourceSource,
             logSource,
         )
-        val payload = source.getEnvelope(SessionSnapshotType.NORMAL_END)
+        val payload = source.getEnvelope()
         assertEquals(metadataSource.metadata, payload.metadata)
         assertEquals(resourceSource.resource, payload.resource)
         assertEquals(logSource.logs, payload.data)
